legg vedtaks mappe som en del av saksbehandling
