+++ src/pages/saksbehandling/søknadsbehandling/sendTilAttesteringPage/sendTilAttesteringPage-nb.ts
