+++ src/pages/saksbehandling/vedtak/søknadsbehandlingvedtakoppsummering/søknadsbehandlingvedtakoppsummering-nb.ts
