+++ src/pages/saksbehandling/vedtak/vedtaksoppsummering.module.less
